backend/requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.91.0
h11==0.14.0
idna==3.4
pydantic==1.10.4
sniffio==1.3.0
starlette==0.24.0
typing-extensions==4.4.0
uvicorn==0.20.0
numpy==1.25.2
sqlalchemy==2.0.20
psycopg2==2.9.7
alembic
pytest
httpx
pytest-asyncio
Authlib
itsdangerous
PyJWT
